Add Red Button One as live channel.
Red Button One is the only one remain of the original red button channels. While special events broadcasts are now often available as web streams that come and go as needed, red button one is a linear stream.
Since programmes are sometimes hard or impossible to find in the normal listings of iplayer, it makes sense to add red button as live stream.